moved kernel functions and tests to separate files
